+++ package/src/dom/nodes/RenderNode.ts
@@ -1,0 +1,168 @@
+import type { RefObject } from "react";
+
+import type {
+  Skia,
+  SkMatrix,
+  SkRect,
+  SkRRect,
+  SkPath,
+  SkPaint,
+} from "../../skia/types";
+import { ClipOp, processTransform, isRRect } from "../../skia/types";
+import type {
+  RenderNode,
+  GroupProps,
+  DrawingContext,
+  NodeType,
+} from "../types";
+
+import { isPathDef, processPath } from "./datatypes";
+import { JsiNode } from "./Node";
+import { PaintNode } from "./PaintNode";
+
+const isSkPaint = (obj: RefObject<SkPaint> | SkPaint): obj is SkPaint =>
+  "__typename__" in obj && obj.__typename__ === "Paint";
+
+export abstract class JsiRenderNode<P extends GroupProps>
+  extends JsiNode<P>
+  implements RenderNode<P>
+{
+  paint?: PaintNode;
+  matrix?: SkMatrix;
+  clipRect?: SkRect;
+  clipRRect?: SkRRect;
+  clipPath?: SkPath;
+
+  constructor(Skia: Skia, type: NodeType, props: P) {
+    super(Skia, type, props);
+    this.onPropChange();
+  }
+
+  setProps(props: P) {
+    super.setProps(props);
+    this.onPropChange();
+  }
+
+  protected onPropChange() {
+    this.matrix = undefined;
+    this.clipPath = undefined;
+    this.clipRect = undefined;
+    this.clipRRect = undefined;
+    this.paint = undefined;
+    if (this.hasCustomPaint()) {
+      this.paint = new PaintNode(this.Skia, this.props);
+    }
+    this.computeMatrix();
+    this.computeClip();
+  }
+
+  private computeClip() {
+    const { clip } = this.props;
+    if (clip) {
+      if (isPathDef(clip)) {
+        this.clipPath = processPath(this.Skia, clip);
+      } else if (isRRect(clip)) {
+        this.clipRRect = clip;
+      } else {
+        this.clipRect = clip;
+      }
+    }
+  }
+
+  private computeMatrix() {
+    const { transform, origin, matrix } = this.props;
+    if (matrix) {
+      this.matrix = matrix;
+    } else if (transform) {
+      const m = this.Skia.Matrix();
+      if (origin) {
+        m.translate(origin.x, origin.y);
+      }
+      processTransform(m, transform);
+      if (origin) {
+        m.translate(-origin.x, -origin.y);
+      }
+      this.matrix = m;
+    }
+  }
+
+  private hasCustomPaint() {
+    const {
+      color,
+      strokeWidth,
+      blendMode,
+      style,
+      strokeJoin,
+      strokeCap,
+      strokeMiter,
+      opacity,
+      antiAlias,
+    } = this.props;
+    return (
+      color !== undefined ||
+      strokeWidth !== undefined ||
+      blendMode !== undefined ||
+      style !== undefined ||
+      strokeJoin !== undefined ||
+      strokeCap !== undefined ||
+      strokeMiter !== undefined ||
+      opacity !== undefined ||
+      antiAlias !== undefined
+    );
+  }
+
+  render(parentCtx: DrawingContext) {
+    const { invertClip, layer } = this.props;
+    const { canvas } = parentCtx;
+
+    const opacity = this.props.opacity
+      ? parentCtx.opacity * this.props.opacity
+      : parentCtx.opacity;
+
+    const paint = this.paint
+      ? this.paint.concat(parentCtx.paint, opacity)
+      : parentCtx.paint;
+
+    // TODO: can we only recreate a new context here if needed?
+    const ctx = { ...parentCtx, opacity, paint };
+    const hasTransform = this.matrix !== undefined;
+    const hasClip = this.clipRect !== undefined;
+    const shouldSave = hasTransform || hasClip || !!layer;
+    const op = invertClip ? ClipOp.Difference : ClipOp.Intersect;
+
+    if (shouldSave) {
+      if (layer) {
+        if (typeof layer === "boolean") {
+          canvas.saveLayer();
+        } else if (isSkPaint(layer)) {
+          canvas.saveLayer(layer);
+        } else {
+          canvas.saveLayer(layer.current ?? undefined);
+        }
+      } else {
+        canvas.save();
+      }
+    }
+
+    if (this.matrix) {
+      canvas.concat(this.matrix);
+    }
+    if (this.clipRect) {
+      canvas.clipRect(this.clipRect, op, true);
+    }
+    if (this.clipRRect) {
+      canvas.clipRRect(this.clipRRect, op, true);
+    }
+    if (this.clipPath) {
+      canvas.clipPath(this.clipPath, op, true);
+    }
+
+    this.renderNode(ctx);
+
+    if (shouldSave) {
+      canvas.restore();
+    }
+  }
+
+  abstract renderNode(ctx: DrawingContext): void;
+}
