+++ package/src/dom/nodes/drawings/CustomDrawingNode.ts
@@ -1,8 +1,7 @@
 import type { Skia } from "../../../skia/types";
 import type { CustomDrawingNodeProps, DrawingContext } from "../../types";
 import { NodeType } from "../../types";
-
-import { JsiDrawingNode } from "./DrawingNode";
+import { JsiDrawingNode } from "../DrawingNode";
 
 export class CustomDrawingNode extends JsiDrawingNode<CustomDrawingNodeProps> {
   constructor(Skia: Skia, props: CustomDrawingNodeProps) {
