+++ package/src/dom/nodes/drawings/DiffRectNode.ts
@@ -1,8 +1,7 @@
 import type { Skia } from "../../../skia/types";
 import type { DiffRectProps, DrawingContext } from "../../types";
 import { NodeType } from "../../types";
-
-import { JsiDrawingNode } from "./DrawingNode";
+import { JsiDrawingNode } from "../DrawingNode";
 
 export class DiffRectNode extends JsiDrawingNode<DiffRectProps> {
   constructor(Skia: Skia, props: DiffRectProps) {
