+++ package/src/dom/nodes/drawings/FillNode.ts
@@ -1,8 +1,7 @@
 import type { Skia } from "../../../skia/types";
 import type { DrawingContext, DrawingNodeProps } from "../../types";
 import { NodeType } from "../../types";
-
-import { JsiDrawingNode } from "./DrawingNode";
+import { JsiDrawingNode } from "../DrawingNode";
 
 export class FillNode extends JsiDrawingNode<DrawingNodeProps> {
   constructor(Skia: Skia, props: DrawingNodeProps = {}) {
