+++ package/src/dom/nodes/drawings/OvalNode.ts
@@ -2,8 +2,7 @@ import type { Skia, SkRect } from "../../../skia/types";
 import type { DrawingContext, OvalProps } from "../../types";
 import { NodeType } from "../../types";
 import { processRect } from "../datatypes";
-
-import { JsiDrawingNode } from "./DrawingNode";
+import { JsiDrawingNode } from "../DrawingNode";
 
 export class OvalNode extends JsiDrawingNode<OvalProps> {
   rect?: SkRect;
