+++ package/src/dom/nodes/drawings/index.ts
@@ -10,7 +10,6 @@ export * from "./LineNode";
 export * from "./PatchNode";
 export * from "./VerticesNode";
 export * from "./CustomDrawingNode";
-export * from "./DrawingNode";
 export * from "./Text";
 export * from "./PictureNode";
 export * from "./DiffRectNode";
